REF: Move greater_is_better from metrics.py to utils.py
