+++ orca_python/metrics/__init__.py
@@ -6,7 +6,6 @@ from .metrics import (
     ccr,
     gm,
     gmsec,
-    greater_is_better,
     mae,
     mmae,
     ms,
@@ -16,9 +15,9 @@ from .metrics import (
     tkendall,
     wkappa,
 )
+from .utils import get_metric_names, greater_is_better
 
 __all__ = [
-    "greater_is_better",
     "ccr",
     "amae",
     "gm",
@@ -32,4 +31,6 @@ __all__ = [
     "spearman",
     "rps",
     "accuracy_off1",
+    "get_metric_names",
+    "greater_is_better",
 ]
